Leave poly2tri behind part 1: upgrade rubber band highlight
