--- src/core/sgrubberband.h
@@ -1,47 +1,0 @@
-/***************************************************************************
-    sgrubberband.h
-    ---------------------
-    begin                : March 2020
-    copyright            : (C) 2020 by David Signer
-    email                : david at opengis dot ch
- ***************************************************************************
- *                                                                         *
- *   This program is free software; you can redistribute it and/or modify  *
- *   it under the terms of the GNU General Public License as published by  *
- *   the Free Software Foundation; either version 2 of the License, or     *
- *   (at your option) any later version.                                   *
- *                                                                         *
- ***************************************************************************/
-#ifndef QGSSGRUBBERBAND_H
-#define QGSSGRUBBERBAND_H
-
-#include <QtQuick/QSGFlatColorMaterial>
-#include <QtQuick/QSGNode>
-#include <qgspoint.h>
-#include <qgswkbtypes.h>
-
-
-class RubberbandModel;
-
-/**
- * This is used to render a rubberband on the scene graph.
- *
- * This cannot be considered stable API.
- */
-
-class SGRubberband : public QSGNode
-{
-  public:
-    SGRubberband( const QVector<QgsPoint> &points, Qgis::GeometryType type, const QColor &color, float width, const QColor &outlineColor, float outlineWidth );
-
-  private:
-    QSGGeometryNode *createLineGeometry( const QVector<QgsPoint> &points );
-    QSGGeometryNode *createPolygonGeometry( const QVector<QgsPoint> &points );
-
-    QSGFlatColorMaterial mMaterial;
-    QSGFlatColorMaterial mOutlineMaterial;
-    float mWidth;
-    float mOutlineWidth;
-};
-
-#endif // QGSSGRUBBERBAND_H
