+++ src/qml/qml.qrc
@@ -67,6 +67,7 @@
         <file>FeatureListSelectionHighlight.qml</file>
         <file>editorwidgets/ValueRelation.qml</file>
         <file>RelationCombobox.qml</file>
+        <file>Rubberband.qml</file>
         <file>Changelog.qml</file>
         <file>Nyuki.qml</file>
         <file>LayerLoginDialog.qml</file>
